refactor(cardano): move delta structs alongside entities
