+++ crates/cardano/src/estart/reset.rs
@@ -1,139 +1,10 @@
-use std::sync::Arc;
-
-use dolos_core::{ChainError, Genesis, NsKey};
-use pallas::ledger::primitives::Epoch;
-use serde::{Deserialize, Serialize};
+use dolos_core::ChainError;
 
 use crate::{
     estart::{AccountId, PoolId, WorkContext},
     pots::{apply_delta, PotDelta, Pots},
-    AccountState, CardanoDelta, EpochState, EraTransition, FixedNamespace as _, PoolState,
-    CURRENT_EPOCH_KEY,
+    AccountState, AccountTransition, CardanoDelta, EpochTransition, PoolState, PoolTransition,
 };
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AccountTransition {
-    account: AccountId,
-    next_epoch: Epoch,
-}
-
-impl AccountTransition {
-    pub fn new(account: AccountId, next_epoch: Epoch) -> Self {
-        Self {
-            account,
-            next_epoch,
-        }
-    }
-}
-
-impl dolos_core::EntityDelta for AccountTransition {
-    type Entity = AccountState;
-
-    fn key(&self) -> NsKey {
-        NsKey::from((AccountState::NS, self.account.clone()))
-    }
-
-    fn apply(&mut self, entity: &mut Option<AccountState>) {
-        let entity = entity.as_mut().expect("existing account");
-
-        // apply changes
-        entity.stake.default_transition(self.next_epoch);
-        entity.pool.default_transition(self.next_epoch);
-        entity.drep.default_transition(self.next_epoch);
-    }
-
-    fn undo(&self, _entity: &mut Option<AccountState>) {
-        // todo!()
-        // Placeholder undo logic. Ensure this does not panic.
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PoolTransition {
-    pool: PoolId,
-    next_epoch: Epoch,
-}
-
-impl PoolTransition {
-    pub fn new(pool: PoolId, next_epoch: Epoch) -> Self {
-        Self { pool, next_epoch }
-    }
-}
-
-impl dolos_core::EntityDelta for PoolTransition {
-    type Entity = PoolState;
-
-    fn key(&self) -> NsKey {
-        NsKey::from((PoolState::NS, self.pool.clone()))
-    }
-
-    fn apply(&mut self, entity: &mut Option<PoolState>) {
-        let entity = entity.as_mut().expect("existing pool");
-
-        // apply changes
-        entity.snapshot.default_transition(self.next_epoch);
-    }
-
-    fn undo(&self, _entity: &mut Option<PoolState>) {
-        // todo!()
-        // Placeholder undo logic. Ensure this does not panic.
-    }
-}
-
-#[derive(Clone, Serialize, Deserialize)]
-pub struct EpochTransition {
-    new_epoch: Epoch,
-    new_pots: Pots,
-    era_transition: Option<EraTransition>,
-
-    #[serde(skip)]
-    genesis: Option<Arc<Genesis>>,
-}
-
-impl std::fmt::Debug for EpochTransition {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "EpochTransition")?;
-        Ok(())
-    }
-}
-
-impl dolos_core::EntityDelta for EpochTransition {
-    type Entity = EpochState;
-
-    fn key(&self) -> NsKey {
-        NsKey::from((EpochState::NS, CURRENT_EPOCH_KEY))
-    }
-
-    fn apply(&mut self, entity: &mut Option<Self::Entity>) {
-        let entity = entity.as_mut().expect("existing epoch");
-
-        debug_assert!(self
-            .new_pots
-            .is_consistent(entity.initial_pots.max_supply()));
-
-        entity.number = self.new_epoch;
-        entity.initial_pots = self.new_pots.clone();
-        entity.rolling.default_transition(self.new_epoch);
-        entity.pparams.default_transition(self.new_epoch);
-
-        // if we have an era transition, we need to migrate the pparams
-        if let Some(transition) = &self.era_transition {
-            let current = entity.pparams.unwrap_live_mut();
-
-            *current = crate::forks::migrate_pparams_version(
-                transition.prev_version.into(),
-                transition.new_version.into(),
-                current,
-                self.genesis.as_ref().expect("genesis not set"),
-            );
-        }
-    }
-
-    fn undo(&self, _entity: &mut Option<Self::Entity>) {
-        // todo!()
-        // Placeholder undo logic. Ensure this does not panic.
-    }
-}
 
 pub fn define_new_pots(ctx: &super::WorkContext) -> Pots {
     let epoch = ctx.ended_state();
