+++ crates/cardano/src/ewrap/refunds.rs
@@ -1,89 +1,11 @@
-use dolos_core::{ChainError, EntityKey, NsKey};
-use pallas::{codec::minicbor, ledger::primitives::StakeCredential};
-use serde::{Deserialize, Serialize};
+use dolos_core::{ChainError, EntityKey};
 
 use crate::{
+    credential_to_key,
     ewrap::{BoundaryWork, ProposalId},
-    rupd::credential_to_key,
-    AccountState, CardanoDelta, CardanoEntity, FixedNamespace as _, PoolDepositRefundLog, PoolHash,
-    PoolState, ProposalState,
+    AccountState, CardanoDelta, CardanoEntity, PoolDepositRefund, PoolDepositRefundLog, PoolHash,
+    PoolState, ProposalDepositRefund, ProposalState,
 };
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PoolDepositRefund {
-    pool_deposit: u64,
-    account: StakeCredential,
-}
-
-impl PoolDepositRefund {
-    pub fn new(pool_deposit: u64, account: StakeCredential) -> Self {
-        Self {
-            pool_deposit,
-            account,
-        }
-    }
-}
-
-impl dolos_core::EntityDelta for PoolDepositRefund {
-    type Entity = AccountState;
-
-    fn key(&self) -> NsKey {
-        let enc = minicbor::to_vec(&self.account).unwrap();
-        NsKey::from((AccountState::NS, enc))
-    }
-
-    fn apply(&mut self, entity: &mut Option<Self::Entity>) {
-        let entity = entity.as_mut().expect("existing account");
-
-        let stake = entity.stake.scheduled_or_default();
-
-        stake.rewards_sum += self.pool_deposit;
-    }
-
-    fn undo(&self, _entity: &mut Option<Self::Entity>) {
-        // todo!()
-        // Placeholder undo logic. Ensure this does not panic.
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ProposalDepositRefund {
-    proposal_deposit: u64,
-    account: StakeCredential,
-}
-
-impl ProposalDepositRefund {
-    pub fn new(proposal_deposit: u64, account: StakeCredential) -> Self {
-        Self {
-            proposal_deposit,
-            account,
-        }
-    }
-}
-
-impl dolos_core::EntityDelta for ProposalDepositRefund {
-    type Entity = AccountState;
-
-    fn key(&self) -> NsKey {
-        let enc = minicbor::to_vec(&self.account).unwrap();
-        NsKey::from((AccountState::NS, enc))
-    }
-
-    fn apply(&mut self, entity: &mut Option<Self::Entity>) {
-        let entity = entity.as_mut().expect("existing account");
-
-        tracing::debug!(cred=?self.account, deposit=%self.proposal_deposit, "applying proposal deposit refund");
-
-        let stake = entity.stake.scheduled_or_default();
-
-        stake.rewards_sum += self.proposal_deposit;
-    }
-
-    fn undo(&self, _entity: &mut Option<Self::Entity>) {
-        // todo!()
-        // Placeholder undo logic. Ensure this does not panic.
-    }
-}
 
 #[derive(Default)]
 pub struct BoundaryVisitor {
