+++ crates/cardano/src/roll/dreps.rs
@@ -1,16 +1,15 @@
 use std::{collections::HashMap, ops::Deref as _};
 
-use dolos_core::{BlockSlot, ChainError, NsKey, TxOrder, TxoRef};
+use dolos_core::{ChainError, TxOrder, TxoRef};
 use pallas::ledger::{
-    primitives::conway::{self, Anchor, DRep, Voter},
+    primitives::conway::{self, DRep, Voter},
     traverse::{MultiEraBlock, MultiEraCert, MultiEraTx},
 };
-use serde::{Deserialize, Serialize};
 
 use super::WorkDeltas;
 use crate::{
-    drep_to_entity_key, model::DRepState, owned::OwnedMultiEraOutput,
-    pallas_extras::stake_cred_to_drep, roll::BlockVisitor, FixedNamespace as _,
+    owned::OwnedMultiEraOutput, pallas_extras::stake_cred_to_drep, roll::BlockVisitor,
+    DRepActivity, DRepRegistration, DRepUnRegistration,
 };
 
 fn cert_drep(cert: &MultiEraCert) -> Option<DRep> {
@@ -22,158 +21,6 @@ fn cert_drep(cert: &MultiEraCert) -> Option<DRep> {
             _ => None,
         },
         _ => None,
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DRepRegistration {
-    drep: DRep,
-    slot: BlockSlot,
-    txorder: TxOrder,
-    deposit: u64,
-    anchor: Option<Anchor>,
-
-    // undo
-    prev_deposit: Option<u64>,
-}
-
-impl DRepRegistration {
-    pub fn new(
-        drep: DRep,
-        slot: BlockSlot,
-        txorder: TxOrder,
-        deposit: u64,
-        anchor: Option<Anchor>,
-    ) -> Self {
-        Self {
-            drep,
-            slot,
-            txorder,
-            deposit,
-            anchor,
-            prev_deposit: None,
-        }
-    }
-}
-
-impl dolos_core::EntityDelta for DRepRegistration {
-    type Entity = DRepState;
-
-    fn key(&self) -> NsKey {
-        NsKey::from((DRepState::NS, drep_to_entity_key(&self.drep)))
-    }
-
-    fn apply(&mut self, entity: &mut Option<DRepState>) {
-        let entity = entity.get_or_insert_with(|| DRepState::new(self.drep.clone()));
-
-        // apply changes
-        entity.registered_at = Some((self.slot, self.txorder));
-        entity.voting_power = self.deposit;
-        entity.deposit = self.deposit;
-    }
-
-    fn undo(&self, entity: &mut Option<DRepState>) {
-        let entity = entity.get_or_insert_with(|| DRepState::new(self.drep.clone()));
-
-        entity.registered_at = None;
-        entity.voting_power = 0;
-        entity.deposit = self.prev_deposit.unwrap_or(0);
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DRepUnRegistration {
-    drep: DRep,
-    slot: BlockSlot,
-    txorder: TxOrder,
-
-    // undo data
-    prev_voting_power: Option<u64>,
-    prev_deposit: Option<u64>,
-    prev_unregistered_at: Option<(BlockSlot, TxOrder)>,
-}
-
-impl DRepUnRegistration {
-    pub fn new(drep: DRep, slot: BlockSlot, txorder: TxOrder) -> Self {
-        Self {
-            drep,
-            slot,
-            txorder,
-            prev_voting_power: None,
-            prev_deposit: None,
-            prev_unregistered_at: None,
-        }
-    }
-}
-
-impl dolos_core::EntityDelta for DRepUnRegistration {
-    type Entity = DRepState;
-
-    fn key(&self) -> NsKey {
-        NsKey::from((DRepState::NS, drep_to_entity_key(&self.drep)))
-    }
-
-    fn apply(&mut self, entity: &mut Option<DRepState>) {
-        let entity = entity.as_mut().expect("can't unregister missing drep");
-
-        // save undo data
-        self.prev_voting_power = Some(entity.voting_power);
-        self.prev_unregistered_at = entity.unregistered_at;
-        self.prev_deposit = Some(entity.deposit);
-
-        // apply changes
-        entity.voting_power = 0;
-        entity.unregistered_at = Some((self.slot, self.txorder));
-        entity.deposit = 0;
-    }
-
-    fn undo(&self, entity: &mut Option<DRepState>) {
-        let entity = entity.as_mut().expect("can't undo missing drep");
-
-        entity.voting_power = self.prev_voting_power.unwrap();
-        entity.unregistered_at = self.prev_unregistered_at;
-        entity.deposit = self.prev_deposit.unwrap_or(0);
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct DRepActivity {
-    drep: DRep,
-    slot: u64,
-    previous_last_active_slot: Option<u64>,
-}
-
-impl DRepActivity {
-    pub fn new(drep: DRep, slot: u64) -> Self {
-        Self {
-            drep,
-            slot,
-            previous_last_active_slot: None,
-        }
-    }
-}
-
-impl dolos_core::EntityDelta for DRepActivity {
-    type Entity = DRepState;
-
-    fn key(&self) -> NsKey {
-        NsKey::from((DRepState::NS, drep_to_entity_key(&self.drep)))
-    }
-
-    fn apply(&mut self, entity: &mut Option<DRepState>) {
-        let entity = entity.get_or_insert_with(|| DRepState::new(self.drep.clone()));
-
-        // save undo info
-        self.previous_last_active_slot = entity.last_active_slot;
-
-        // apply changes
-        entity.last_active_slot = Some(self.slot);
-    }
-
-    fn undo(&self, entity: &mut Option<DRepState>) {
-        let entity = entity.as_mut().expect("can't undo missing drep");
-
-        entity.last_active_slot = self.previous_last_active_slot;
     }
 }
